Refactor forms - part 2
