+++ src/lib/components/forms/email-form.svelte
@@ -9,9 +9,13 @@
 
 	import FormWrapper from './form-wrapper.svelte';
 
-	export let data: SuperValidated<Infer<EmailFormSchema>>;
+	type Props = {
+		data: SuperValidated<Infer<EmailFormSchema>>;
+		buttonLabel?: string;
+		action: string;
+	};
 
-	export let buttonLabel: string = m.few_blue_wallaby_read();
+	let { action, data, buttonLabel = m.few_blue_wallaby_read() }: Props = $props();
 
 	const form = superForm(data, {
 		validators: zodClient(emailFormSchema()),
@@ -30,7 +34,7 @@
 </script>
 
 <FormWrapper message={$message}>
-	<form method="POST" use:enhance>
+	<form method="POST" use:enhance {action}>
 		<Form.Field {form} name="email" class="py-4">
 			<Form.Control let:attrs>
 				<Form.Label>{m.clear_lost_goose_beam()}</Form.Label>
