+++ src/routes/(minimal)/login/+page.svelte
@@ -13,7 +13,7 @@
 </script>
 
 <SingleFormPage title={m.legal_weak_jay_bless()} description={m.ago_crazy_pelican_pray()}>
-	<EmailForm data={data.form} />
+	<EmailForm data={data.emailForm} action="?/loginWithEmail" />
 	<div class="px-5 text-center text-sm">
 		{m.warm_ideal_butterfly_radiate()}
 		<Link href="/signup">{m.grassy_sea_pug_gasp()}</Link>
